refactor: P3-T14 - 拆分600行以上大文件
